Mon, Mar 20, 2023, 8:41 PM +00:00

+++ 0/aula4.c
@@ -4,10 +4,10 @@
 
 int main(){
     setlocale(LC_ALL, "Portuguese");
-    int n;
+    float  n;
 
     printf("Insira o valor: ");
-    scanf("%d", &n);
+    scanf("%f", &n);
 
     if(n >= '20')
     {
